@@ -11,24 +11,23 @@
   "dependencies": {
     "@vueuse/core": "^9.13.0",
     "dayjs": "^1.11.7",
-    "element-plus": "^2.3.3",
+    "element-plus": "^2.3.4",
     "lodash-es": "^4.17.21",
     "mitt": "^3.0.0",
     "nprogress": "^0.2.0",
-    "pinia": "^2.0.33",
+    "pinia": "^2.0.36",
     "pinia-persistedstate-plugin": "^0.1.0",
     "pinia-plugin-persistedstate": "^3.1.0",
     "sortablejs": "^1.15.0",
-    "vite-plugin-svg-icons": "^2.0.1",
-    "vue": "^3.2.25",
-    "vue-router": "^4.1.6"
+    "vue": "^3.3.2",
+    "vue-router": "^4.2.0"
   },
   "devDependencies": {
     "@types/node": "^18.15.3",
     "@types/nprogress": "^0.2.0",
     "@typescript-eslint/eslint-plugin": "^5.55.0",
-    "@vitejs/plugin-vue": "^2.3.4",
-    "@vitejs/plugin-vue-jsx": "^1.3.10",
+    "@vitejs/plugin-vue": "^4.2.3",
+    "@vitejs/plugin-vue-jsx": "^3.0.1",
     "eslint": "^8.36.0",
     "eslint-config-prettier": "^8.7.0",
     "eslint-define-config": "^1.17.0",
@@ -36,12 +35,28 @@
     "eslint-plugin-vue": "^9.9.0",
     "less": "^4.1.3",
     "prettier": "^2.8.4",
-    "typescript": "^5.0.2",
+    "rollup-plugin-visualizer": "^5.9.0",
+    "typescript": "^5.0.4",
     "unplugin-auto-import": "^0.15.1",
     "unplugin-vue-components": "^0.24.1",
-    "vite": "^2.9.15",
+    "unplugin-vue-setup-extend-plus": "^1.0.0",
+    "vite": "^4.3.7",
     "vite-plugin-eslint": "^1.8.1",
-    "vite-plugin-vue-setup-extend": "^0.4.0",
-    "vue-tsc": "^0.35.0"
+    "vite-plugin-svg-icons": "^2.0.1",
+    "vite-plugin-html": "^3.2.0",
+    "vite-plugin-compression": "^0.5.1",
+    "vue-tsc": "^1.6.5"
+  },
+  "browserslist": {
+    "production": [
+      "> 1%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
